Add AudioFile semantic type
