+++ autogoal/kb/_semantics.py
@@ -276,18 +276,6 @@ class Seq(SemanticType):
         return SeqImp
 
 
-from os import path
-class AudioFile(SemanticType):
-    """Semantic type for wav audio files"""
-
-    @classmethod
-    def _match(cls, x):
-        try:
-            return path.isfile(x) and x.split(".")[-1] == "wav"
-        except TypeError:
-            return False
-
-
 # Now let's move to the algebraic types, vectors, matrices, and tensors.
 # These wrap numpy arrays of different dimensionalities
 # We'll have three different semantic labels for each tensor: dimensionality (an integer),
